ARM: dts: dumpling: Don't slowly ramp brightness for hbm on/off
The hbm off commands are executed on panel power on by the LiveDisplay
driver when the panel is turned on and hbm is disabled. However, since
the current hbm command causes brightness to slowly ramp up/down, this
results in the screen power on using a slow brightness ramp up/down which
isn't preferred. The slow brightness ramp when enabling hbm isn't wanted
either, so disable the brightness ramping for both hbm on and off commands.

Change-Id: I0b1be3c7dcb02d349c2b0b469080589eeab235bf
Signed-off-by: Sultanxda <sultanxda@gmail.com>