Update patch set 1
Patch Set 1: Code-Review+1

The problem was introduced by 6373b355 "sound: usb: Enable autosuspend for usb audio device" which was not merged in mainline kernel. We should revert that commit instead (or check if our USB controller code is broken) if we could confirm this is a common problem for audio peripherals.

Patch-set: 1
Label: Code-Review=+1